# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/campaigns/approve": {
            "patch": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can Approve Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Campaign"
                ],
                "summary": "Admin can Approv Campaign",
                "parameters": [
                    {
                        "description": "Post ID ",
                        "name": "approveCampaignBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.ApproveCampaignBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.ApproveCampaignResponse"
                        }
                    }
                }
            }
        },
        "/admin/campaigns/details": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can get Campaign Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Campaign"
                ],
                "summary": "Admin can get Campaign Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CampaignDetailsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/campaigns/reported/delete": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can Delete reported post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Reported"
                ],
                "summary": "Admin can Delete reported post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID ",
                        "name": "postId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.DeleteReportResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/campaigns/requestlist": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see campaogn requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Campaign"
                ],
                "summary": "Admin can see campaign request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CampaignRequestListResponse"
                        }
                    }
                }
            }
        },
        "/admin/categories/categorylist": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see Categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin can see Categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CategoryListResponse"
                        }
                    }
                }
            }
        },
        "/admin/categories/categorylist/posts": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see Category posts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Admin can see Category posts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CategoryPostsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/dashboard": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see website statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Admin can see website statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.AdminDashboardResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/post/delete": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can delete post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Feeds"
                ],
                "summary": "Admin can delete post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.DeletePostResponse"
                        }
                    }
                }
            }
        },
        "/admin/post/details": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can get post detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Feeds"
                ],
                "summary": "Admin can get post detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "post id Data",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.PostDetailsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "pb.AdminDashboardResponse": {
            "type": "object",
            "properties": {
                "alltime": {
                    "$ref": "#/definitions/pb.Stats"
                },
                "annual": {
                    "$ref": "#/definitions/pb.Stats"
                },
                "daily": {
                    "$ref": "#/definitions/pb.Stats"
                },
                "monthly": {
                    "$ref": "#/definitions/pb.Stats"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "weekly": {
                    "$ref": "#/definitions/pb.Stats"
                }
            }
        },
...
        "pb.ApproveCampaignResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/pkg_admin_pb.Post"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.CampaignDetailsResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/pkg_admin_pb.Post"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.CampaignRequestListResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_admin_pb.Post"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.Category": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "categoryid": {
                    "type": "integer"
                }
            }
        },
        "pb.CategoryListResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Category"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.CategoryPostsResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/pb.Category"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_admin_pb.Post"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.DeletePostResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.DeleteReportResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.PostDetailsResponse": {
            "type": "object",
            "properties": {
                "Post": {
                    "$ref": "#/definitions/pkg_admin_pb.Post"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.Stats": {
            "type": "object",
            "properties": {
                "activecampaigns": {
                    "type": "integer"
                },
                "collectedmoney": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "posts": {
                    "type": "integer"
                },
                "users": {
                    "type": "integer"
                }
            }
        },
...
        "pkg_admin_pb.Post": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "collected": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
        "routes.ApproveCampaignBody": {
            "type": "object",
            "required": [
                "postId"
            ],
            "properties": {
                "postId": {
                    "type": "integer",
                    "maximum": 999
                }
            }
        },
...
    }
}